+++ mockserver/mockserver-core/src/main/java/org/mockserver/mock/ResponseMode.java
@@ -3,5 +3,6 @@ package org.mockserver.mock;
 public enum ResponseMode {
     SEQUENTIAL,
     RANDOM,
-    WEIGHTED
+    WEIGHTED,
+    SWITCH
 }